directives_meta.py: more use of immutable specs
* Both `@when(...)` and `with when(...)` now use immutable Specs
* Avoid the `Spec -> str -> Spec` conversion between `when` and
  `directives_meta`
* Avoid copy of `kwargs` in the common case where there is no
  `default_args` set
* Remove vestigial `if isinstance(result, Sequence)` check on the return
  value of a directive -- it's always a callable.
* Avoid `constrain` call if there's only a single when condition.
